+++ R/SOS-methods-accessor.R
@@ -34,7 +34,7 @@ if (!isGeneric("sosCaps"))
     standardGeneric("sosCaps")
   })
 setMethod(f = "sosCaps", signature = signature(sos = "SOS"),
-          def = function(sos) {
+          definition = function(sos) {
             return(sos@capabilities)
           })
 
@@ -43,7 +43,7 @@ if (!isGeneric("sosFilter_Capabilities"))
     standardGeneric("sosFilter_Capabilities")
   })
 setMethod(f = "sosFilter_Capabilities", signature = signature(sos = "SOS"),
-          def = function(sos) {
+          definition = function(sos) {
             return(sos@capabilities@filterCapabilities)
           })
 
@@ -52,7 +52,7 @@ if (!isGeneric("sosServiceIdentification"))
     standardGeneric("sosServiceIdentification")
   })
 setMethod(f = "sosServiceIdentification", signature = signature(sos = "SOS"),
-          def = function(sos) {
+          definition = function(sos) {
             return(sos@capabilities@identification)
           })
 
@@ -61,7 +61,7 @@ if (!isGeneric("sosServiceProvider"))
     standardGeneric("sosServiceProvider")
   })
 setMethod(f = "sosServiceProvider", signature = signature(sos = "SOS"),
-          def = function(sos) {
+          definition = function(sos) {
             return(sos@capabilities@provider)
           })
 
@@ -70,7 +70,7 @@ if (!isGeneric("sosOperationsMetadata"))
     standardGeneric("sosOperationsMetadata")
   })
 setMethod(f = "sosOperationsMetadata", signature = signature(sos = "SOS"),
-          def = function(sos) {
+          definition = function(sos) {
             return(sos@capabilities@operations)
           })
 
@@ -79,19 +79,19 @@ if (!isGeneric("sosOperations"))
     standardGeneric("sosOperations")
   })
 setMethod(f = "sosOperations", signature = signature(obj = "SOS"),
-          def = function(obj) {
+          definition = function(obj) {
             return(sosOperations(obj@capabilities))
           })
 setMethod(f = "sosOperations",
           signature = signature(obj = "SosCapabilities_1.0.0"),
-          def = function(obj) {
+          definition = function(obj) {
             if(!is.null(obj@operations))
               return(obj@operations@operations)
             return(NA_character_)
           })
 setMethod(f = "sosOperations",
           signature = signature(obj = "SosCapabilities_2.0.0"),
-          def = function(obj) {
+          definition = function(obj) {
             if(!is.null(obj@operations))
               return(obj@operations@operations)
             return(NA_character_)
@@ -99,7 +99,7 @@ setMethod(f = "sosOperations",
 # required to handle the first capabilities request:
 setMethod(f = "sosOperations",
           signature = signature(obj = "OwsCapabilities"),
-          def = function(obj) {
+          definition = function(obj) {
             return(NULL)
           })
 
@@ -108,7 +108,7 @@ if (!isGeneric("sosContents"))
     standardGeneric("sosContents")
   })
 setMethod(f = "sosContents", signature = signature(sos = "SOS"),
-          def = function(sos) {
+          definition = function(sos) {
             return(sosCaps(sos)@contents)
           })
 
@@ -117,11 +117,11 @@ if (!isGeneric("sosUrl"))
     standardGeneric("sosUrl")
   })
 setMethod(f = "sosUrl", signature = signature(sos = "SOS_1.0.0"),
-          def = function(sos) {
+          definition = function(sos) {
             return(sos@url)
           })
 setMethod(f = "sosUrl", signature = signature(sos = "SOS_2.0.0"),
-          def = function(sos) {
+          definition = function(sos) {
             return(sos@url)
           })
 
@@ -130,7 +130,7 @@ if (!isGeneric("sosVersion"))
     standardGeneric("sosVersion")
   })
 setMethod(f = "sosVersion", signature = signature(sos = "SOS"),
-          def = function(sos) {
+          definition = function(sos) {
             return(sos@version)
           })
 if (!isGeneric("sosSwitchCoordinates"))
@@ -138,7 +138,7 @@ if (!isGeneric("sosSwitchCoordinates"))
     standardGeneric("sosSwitchCoordinates")
   })
 setMethod(f = "sosSwitchCoordinates", signature = signature(sos = "SOS"),
-          def = function(sos) {
+          definition = function(sos) {
             return(sos@switchCoordinates)
           })
 
@@ -148,11 +148,11 @@ if (!isGeneric("sosBinding"))
     standardGeneric("sosBinding")
   })
 setMethod(f = "sosBinding", signature = signature(sos = "SOS_1.0.0"),
-          def = function(sos) {
+          definition = function(sos) {
             return(sos@binding)
           })
 setMethod(f = "sosBinding", signature = signature(sos = "SOS_2.0.0"),
-          def = function(sos) {
+          definition = function(sos) {
             return(sos@binding)
           })
 
@@ -161,56 +161,61 @@ if (!isGeneric("sosProcedures"))
     standardGeneric("sosProcedures")
   })
 setMethod(f = "sosProcedures", signature = signature(obj = "SOS"),
-          def = function(obj) {
+          definition = function(obj) {
             .offerings <- sosOfferings(obj)
             if(length(.offerings) == 1 && is.na(.offerings))
               return(NA_character_)
 
-            .p <- lapply(.offerings, sosProcedures)
+            .p <- sapply(.offerings, sosProcedures)
             names(.p) <- names(.offerings)
             return(.p)
           })
 setMethod(f = "sosProcedures",
           signature = signature(obj = "SosObservationOffering"),
-          def = function(obj) {
+          definition = function(obj) {
             .p <- as.character(obj@procedure)
             return(.p)
           })
 setMethod(f = "sosProcedures",
           signature = signature(obj = "SosObservationOffering_2.0.0"),
-          def = function(obj) {
+          definition = function(obj) {
             .p <- as.character(obj@procedure)
             return(.p)
           })
 setMethod(f = "sosProcedures",
           signature = signature(obj = "list"),
-          def = function(obj) {
+          definition = function(obj) {
             .p <- sapply(obj, sosProcedures)
             return(.p)
           })
 setMethod(f = "sosProcedures",
           signature = signature(obj = "OmObservationCollection"),
-          def = function(obj) {
+          definition = function(obj) {
             .p <- sapply(obj@members, sosProcedures)
             return(.p)
           })
 setMethod(f = "sosProcedures",
           signature = signature(obj = "OmObservation"),
-          def = function(obj) {
+          definition = function(obj) {
             .p <- as.character(obj@procedure)
             return(.p)
           })
 setMethod(f = "sosProcedures",
           signature = signature(obj = "OmOM_Observation"),
-          def = function(obj) {
+          definition = function(obj) {
             .p <- as.character(obj@procedure)
             return(.p)
           })
 setMethod(f = "sosProcedures",
           signature = signature(obj = "OmMeasurement"),
-          def = function(obj) {
+          definition = function(obj) {
             .p <- as.character(obj@procedure)
             return(.p)
+          })
+setMethod(f = "sosProcedures",
+          signature = signature(obj = "DataAvailabilityMember"),
+          definition = function(obj) {
+            return(obj@procedure)
           })
 
 if (!isGeneric("sosObservedProperties"))
@@ -223,7 +228,7 @@ setMethod(f = "sosObservedProperties", signature = signature(obj = "SOS"),
             if(length(.offerings) == 1 && is.na(.offerings))
               return(NA_character_)
 
-            .op <- lapply(.offerings, sosObservedProperties)
+            .op <- sapply(.offerings, sosObservedProperties)
             return(.op)
           })
 setMethod(f = "sosObservedProperties", signature = signature(obj = "SosObservationOffering"),
@@ -233,19 +238,17 @@ setMethod(f = "sosObservedProperties", signature = signature(obj = "SosObservati
             })
 setMethod(f = "sosObservedProperties", signature = signature(obj = "OmObservationCollection"),
           definition = function(obj) {
-            .op <- lapply(obj@members, sosObservedProperties)
-            #			if(removeDuplicates)
-            #				.op <- unique(.op)[[1]]
+            .op <- sapply(obj@members, sosObservedProperties)
             return(.op)
           })
 setMethod(f = "sosObservedProperties", signature = signature(obj = "list"),
           definition = function(obj) {
-            .op <- lapply(obj, sosObservedProperties)
+            .op <- sapply(obj, sosObservedProperties)
             return(.op)
           })
 setMethod(f = "sosObservedProperties", signature = signature(obj = "OmObservation"),
           definition = function(obj) {
-            if(is.null(obj@observedProperty))
+            if (is.null(obj@observedProperty))
               return(NULL)
 
             .op <- sosObservedProperties(obj@observedProperty)
@@ -261,7 +264,7 @@ setMethod(f = "sosObservedProperties", signature = signature(obj = "OmOM_Observa
           })
 setMethod(f = "sosObservedProperties", signature = signature(obj = "SwePhenomenonProperty"),
           definition = function(obj) {
-            if(!is.na(obj@href)) {
+            if (!is.na(obj@href)) {
               return(obj@href)
             }
             else {
@@ -278,52 +281,49 @@ setMethod(f = "sosObservedProperties", signature = signature(obj = "SwePhenomeno
           definition = function(obj) {
             return(obj@href)
           })
+setMethod(f = "sosObservedProperties", signature = signature(obj = "DataAvailabilityMember"),
+          definition = function(obj) {
+            return(obj@observedProperty)
+          })
 
 if (!isGeneric("sosObservableProperties"))
   setGeneric(name = "sosObservableProperties", def = function(obj, ...) {
     standardGeneric("sosObservableProperties")
   })
 setMethod(f = "sosObservableProperties", signature = signature(obj = "SOS"),
-          def = function(obj) {
+          definition = function(obj) {
             .offerings <- sosOfferings(obj)
-            if(length(.offerings) == 1 && is.na(.offerings))
+            if (length(.offerings) == 1 && is.na(.offerings))
               return(NA_character_)
 
-            .op <- lapply(.offerings, sosObservableProperties)
+            .op <- sapply(.offerings, sosObservableProperties)
             return(.op)
           })
-setMethod(f = "sosObservableProperties", signature = signature(
-  obj = "SosObservationOffering_2.0.0"),
-  def = function(obj) {
+setMethod(f = "sosObservableProperties", signature = signature(obj = "SosObservationOffering_2.0.0"),
+  definition = function(obj) {
     .op <- obj@observableProperty
     return(.op)
   })
-setMethod(f = "sosObservableProperties", signature = signature(
-  obj = "OmObservationCollection"),
-  def = function(obj) {
-    .op <- lapply(obj@members, sosObservableProperties)
-    #			if(removeDuplicates)
-    #				.op <- unique(.op)[[1]]
+setMethod(f = "sosObservableProperties", signature = signature(obj = "OmObservationCollection"),
+  definition = function(obj) {
+    .op <- sapply(obj@members, sosObservableProperties)
     return(.op)
   })
-setMethod(f = "sosObservableProperties", signature = signature(
-  obj = "list"),
-  def = function(obj) {
-    .op <- lapply(obj, sosObservableProperties)
+setMethod(f = "sosObservableProperties", signature = signature(obj = "list"),
+  definition = function(obj) {
+    .op <- sapply(obj, sosObservableProperties)
     return(.op)
   })
-setMethod(f = "sosObservableProperties", signature = signature(
-  obj = "OmObservation"),
-  def = function(obj) {
-    if(is.null(obj@observableProperty))
+setMethod(f = "sosObservableProperties", signature = signature(obj = "OmObservation"),
+  definition = function(obj) {
+    if (is.null(obj@observableProperty))
       return(NULL)
 
     .op <- sosObservableProperties(obj@observedProperty)
     return(.op)
   })
-setMethod(f = "sosObservableProperties", signature = signature(
-  obj = "SwePhenomenonProperty"),
-  def = function(obj) {
+setMethod(f = "sosObservableProperties", signature = signature(obj = "SwePhenomenonProperty"),
+  definition = function(obj) {
     if(!is.na(obj@href)) {
       return(obj@href)
     }
@@ -349,25 +349,25 @@ if (!isGeneric("sosBoundedBy"))
   })
 setMethod(f = "sosBoundedBy",
           signature = signature(obj = "SosObservationOffering"),
-          definition = function(obj, bbox = FALSE) {
+definition = function(obj, bbox = FALSE) {
             boundedBy <- .boundedBy(obj, bbox)
             return(boundedBy)
           })
 setMethod(f = "sosBoundedBy",
           signature = signature(obj = "SosObservationOffering_2.0.0"),
-          definition = function(obj, bbox = FALSE) {
+definition = function(obj, bbox = FALSE) {
             boundedBy <- .boundedBy(obj, bbox)
             return(boundedBy)
           })
 setMethod(f = "sosBoundedBy",
           signature = signature(obj = "list"),
-          definition = function(obj, bbox = FALSE) {
+definition = function(obj, bbox = FALSE) {
             .bb <- lapply(obj, sosBoundedBy, bbox = bbox)
             return(.bb)
           })
 setMethod(f = "sosBoundedBy",
           signature = signature(obj = "OmObservationCollection"),
-          definition = function(obj, bbox = FALSE) {
+definition = function(obj, bbox = FALSE) {
             boundedBy <- .boundedBy(obj, bbox)
             return(boundedBy)
           })
@@ -423,7 +423,7 @@ if (!isGeneric("sosOfferings"))
     standardGeneric("sosOfferings")
   })
 setMethod(f = "sosOfferings", signature = signature(obj = "SOS"),
-          def = function(obj, offeringIDs = c(), name = NA) {
+definition = function(obj, offeringIDs = c(), name = NA) {
             .contents <- sosContents(obj)
             if(is.null(.contents))
               return(NA_character_)
@@ -447,7 +447,7 @@ if (!isGeneric("sosOfferingIds"))
     standardGeneric("sosOfferingIds")
   })
 setMethod(f = "sosOfferingIds", signature = signature(sos = "SOS"),
-          def = function(sos) {
+          definition = function(sos) {
             .offerings <- sosOfferings(sos)
             #			if(length(.offerings) == 1 && !is.na(.offerings))
             return(names(.offerings))
@@ -459,7 +459,7 @@ if (!isGeneric("sosFeaturesOfInterest"))
     standardGeneric("sosFeaturesOfInterest")
   })
 setMethod(f = "sosFeaturesOfInterest", signature = signature(obj = "SOS"),
-          def = function(obj, offerings = sosOfferingIds(obj)) {
+definition = function(obj, offerings = sosOfferingIds(obj)) {
             # via observation offering
             .offerings <- sosOfferings(obj)
             .offerings <- .offerings[offerings]
@@ -469,12 +469,12 @@ setMethod(f = "sosFeaturesOfInterest", signature = signature(obj = "SOS"),
           })
 setMethod(f = "sosFeaturesOfInterest",
           signature = signature(obj = "SosObservationOffering"),
-          def = function(obj) {
+          definition = function(obj) {
             return(obj@featureOfInterest)
           })
 setMethod(f = "sosFeaturesOfInterest",
           signature = signature(obj = "OmObservation"),
-          def = function(obj) {
+          definition = function(obj) {
             .foi <- obj@featureOfInterest
             if(is.list(.foi) && length(.foi) == 1)
               return(.foi[[1]])
@@ -482,7 +482,7 @@ setMethod(f = "sosFeaturesOfInterest",
           })
 setMethod(f = "sosFeaturesOfInterest",
           signature = signature(obj = "OmOM_Observation"),
-          def = function(obj) {
+          definition = function(obj) {
             .foi <- obj@featureOfInterest
             if(is.list(.foi) && length(.foi) == 1)
               return(.foi[[1]])
@@ -490,7 +490,7 @@ setMethod(f = "sosFeaturesOfInterest",
           })
 setMethod(f = "sosFeaturesOfInterest",
           signature = signature(obj = "OmMeasurement"),
-          def = function(obj) {
+          definition = function(obj) {
             .foi <- obj@featureOfInterest
             if(is.list(.foi) && length(.foi) == 1)
               return(.foi[[1]])
@@ -498,21 +498,26 @@ setMethod(f = "sosFeaturesOfInterest",
           })
 setMethod(f = "sosFeaturesOfInterest",
           signature = signature(obj = "OmObservationCollection"),
-          def = function(obj) {
-            .fois <- sapply(obj@members, sosFeaturesOfInterest)
+          definition = function(obj) {
+            .fois <- lapply(obj@members, sosFeaturesOfInterest)
             return(.fois)
           })
 setMethod(f = "sosFeaturesOfInterest",
           signature = signature(obj = "list"),
-          def = function(obj) {
-            .fois <- sapply(obj, sosFeaturesOfInterest)
+          definition = function(obj) {
+            .fois <- lapply(obj, sosFeaturesOfInterest)
             return(.fois)
           })
 setMethod(f = "sosFeaturesOfInterest",
           signature = signature(obj = "GmlFeatureCollection"),
-          def = function(obj) {
-            .fois <- sapply(obj@featureCollection, sosFeaturesOfInterest)
+          definition = function(obj) {
+            .fois <- lapply(obj@featureCollection, sosFeaturesOfInterest)
             return(.fois)
+          })
+setMethod(f = "sosFeaturesOfInterest",
+          signature = signature(obj = "DataAvailabilityMember"),
+          definition = function(obj) {
+            return(obj@featureOfInterest)
           })
 
 if (!isGeneric("sosFeatureIds"))
@@ -522,7 +527,7 @@ if (!isGeneric("sosFeatureIds"))
 setMethod(f = "sosFeatureIds",
           signature = signature(obj = "list"),
           definition = function(obj) {
-            .fois <- lapply(obj, sosFeatureIds)
+            .fois <- sapply(obj, sosFeatureIds)
             return(.fois)
           })
 setMethod(f = "sosFeatureIds",
@@ -588,7 +593,7 @@ if (!isGeneric("sosOperation"))
   })
 setMethod(f = "sosOperation",
           signature = signature(sos = "SOS", operationName = "character"),
-          def = function(sos, operationName) {
+          definition = function(sos, operationName) {
             .caps <- sosCaps(sos)
             return(.caps@operations@operations[[operationName]])
           })
@@ -598,7 +603,7 @@ if (!isGeneric("sosResponseFormats"))
     standardGeneric("sosResponseFormats")
   })
 setMethod(f = "sosResponseFormats", signature = signature(obj = "SOS"),
-          def = function(obj, unique = FALSE, removeNulls = TRUE) {
+definition = function(obj, unique = FALSE, removeNulls = TRUE) {
             #			.caps <- sosCaps(obj)
             #			.getOb <- .caps@operations@operations[[sosGetObservationName]]
             #			return(.getOb@parameters$responseFormat)
@@ -614,17 +619,17 @@ setMethod(f = "sosResponseFormats", signature = signature(obj = "SOS"),
           })
 setMethod(f = "sosResponseFormats",
           signature = signature(obj = "SosObservationOffering"),
-          def = function(obj) {
+          definition = function(obj) {
             return(obj@responseFormat)
           })
 setMethod(f = "sosResponseFormats",
           signature = signature(obj = "SosObservationOffering_2.0.0"),
-          def = function(obj) {
+          definition = function(obj) {
             return(obj@responseFormat)
           })
 setMethod(f = "sosResponseFormats",
           signature = signature(obj = "OwsOperation"),
-          def = function(obj) {
+          definition = function(obj) {
             return(obj@parameters$responseFormat)
           })
 
@@ -633,7 +638,7 @@ if (!isGeneric("sosOutputFormats"))
     standardGeneric("sosOutputFormats")
   })
 setMethod(f = "sosOutputFormats", signature = signature(obj = "SOS"),
-          def = function(obj, unique = FALSE, removeNulls = TRUE) {
+definition = function(obj, unique = FALSE, removeNulls = TRUE) {
             .rf <- sapply(sosOperations(obj), sosOutputFormats)
             if(unique) {
               .c <- do.call(c, .rf)
@@ -646,7 +651,7 @@ setMethod(f = "sosOutputFormats", signature = signature(obj = "SOS"),
           })
 setMethod(f = "sosOutputFormats",
           signature = signature(obj = "OwsOperation"),
-          def = function(obj) {
+          definition = function(obj) {
             return(obj@parameters$outputFormat)
           })
 
@@ -655,7 +660,7 @@ if (!isGeneric("sosResponseMode"))
     standardGeneric("sosResponseMode")
   })
 setMethod(f = "sosResponseMode", signature = signature(obj = "SOS"),
-          def = function(obj, unique = FALSE) {
+definition = function(obj, unique = FALSE) {
             #			.caps <- sosCaps(obj)
             #			.getOb <- .caps@operations@operations[[sosGetObservationName]]
             #			return(.getOb@parameters$responseMode)
@@ -668,12 +673,12 @@ setMethod(f = "sosResponseMode", signature = signature(obj = "SOS"),
           })
 setMethod(f = "sosResponseMode",
           signature = signature(obj = "SosObservationOffering"),
-          def = function(obj) {
+          definition = function(obj) {
             return(obj@responseMode)
           })
 setMethod(f = "sosResponseMode",
           signature = signature(obj = "OwsOperation"),
-          def = function(obj) {
+          definition = function(obj) {
             return(obj@parameters$responseMode)
           })
 
@@ -682,7 +687,7 @@ if (!isGeneric("sosResultModels"))
     standardGeneric("sosResultModels")
   })
 setMethod(f = "sosResultModels", signature = signature(obj = "SOS"),
-          def = function(obj, unique = FALSE) {
+definition = function(obj, unique = FALSE) {
             #			.caps <- sosCaps(obj)
             #			.getOb <- .caps@operations@operations[[sosGetObservationName]]
             #			return(.getOb@parameters$resultModel)
@@ -695,12 +700,12 @@ setMethod(f = "sosResultModels", signature = signature(obj = "SOS"),
           })
 setMethod(f = "sosResultModels",
           signature = signature(obj = "SosObservationOffering"),
-          def = function(obj) {
+          definition = function(obj) {
             return(obj@resultModel)
           })
 setMethod(f = "sosResultModels",
           signature = signature(obj = "OwsOperation"),
-          def = function(obj) {
+          definition = function(obj) {
             return(obj@parameters$resultModel)
           })
 
@@ -709,12 +714,12 @@ if (!isGeneric("sosTime"))
     standardGeneric("sosTime")
   })
 setMethod(f = "sosTime", signature = signature(obj = "SOS"),
-          definition = function(obj, convert = FALSE) {
+definition = function(obj, convert = FALSE) {
             times <- lapply(sosOfferings(obj), sosTime, convert = convert)
             return(times)
           })
 setMethod(f = "sosTime", signature = signature(obj = "SosObservationOffering"),
-          def = function(obj, convert = FALSE) {
+definition = function(obj, convert = FALSE) {
             if (!convert)
               return(obj@time)
 
@@ -727,7 +732,7 @@ setMethod(f = "sosTime", signature = signature(obj = "SosObservationOffering"),
             return(obj@time)
           })
 setMethod(f = "sosTime", signature = signature(obj = "GmlTimePeriod"),
-          def = function(obj, ...) {
+definition = function(obj, ...) {
             .start <- NA
             .end <- NA
 
@@ -746,7 +751,7 @@ setMethod(f = "sosTime", signature = signature(obj = "GmlTimePeriod"),
             return(.period)
           })
 setMethod(f = "sosTime", signature = signature(obj = "GmlTimePosition"),
-          def = function(obj, ...) {
+definition = function(obj, ...) {
             .time <- obj@time
             if (!is.na(obj@indeterminatePosition) && obj@indeterminatePosition == "now") {
               .time <- Sys.time()
@@ -761,7 +766,7 @@ setMethod(f = "sosTime", signature = signature(obj = "GmlTimePosition"),
             return(.time)
           })
 setMethod(f = "sosTime", signature = signature(obj = "GmlTimeInstantProperty"),
-          def = function(obj, ...) {
+definition = function(obj, ...) {
             if(is.na(obj@href))
               return(obj@href)
 
@@ -771,21 +776,21 @@ setMethod(f = "sosTime", signature = signature(obj = "GmlTimeInstantProperty"),
             return(NA)
           })
 setMethod(f = "sosTime", signature = signature(obj = "GmlTimeInstant"),
-          def = function(obj, ...) {
+definition = function(obj, ...) {
             return(sosTime(obj@timePosition))
           })
 setMethod(f = "sosTime", signature = signature(obj = "list"),
-          def = function(obj, convert = FALSE) {
+definition = function(obj, convert = FALSE) {
             return(lapply(X = obj, FUN = sosTime, convert = convert))
           })
 setMethod(f = "sosTime", signature = signature(obj = "OmOM_Observation"),
-          def = function(obj, convert = FALSE) {
+definition = function(obj, convert = FALSE) {
             times <- list(resultTime = sosTime(obj@resultTime),
                           phenomenonTime = sosTime(obj@phenomenonTime))
             return(times)
           })
 setMethod(f = "sosTime", signature = signature(obj = "DataAvailabilityMember"),
-          def = function(obj, ...) {
+definition = function(obj, ...) {
             return(sosTime(obj@phenomenonTime))
           })
 
@@ -794,7 +799,7 @@ if (!isGeneric("sosTimeFormat"))
     standardGeneric("sosTimeFormat")
   })
 setMethod(f = "sosTimeFormat", signature = signature(sos = "SOS"),
-          def = function(sos) {
+          definition = function(sos) {
             return(sos@timeFormat)
           })
 
@@ -803,11 +808,11 @@ if (!isGeneric("sosParsers"))
     standardGeneric("sosParsers")
   })
 setMethod(f = "sosParsers", signature = signature(sos = "SOS"),
-          def = function(sos) {
+          definition = function(sos) {
             return(sos@parsers)
           })
 setMethod(f = "sosParsers", signature = signature(sos = "SOS_Test"),
-          def = function(sos) {
+          definition = function(sos) {
             return(sos@parsers)
           })
 
@@ -817,7 +822,7 @@ if (!isGeneric("sosResult"))
     standardGeneric("sosResult")
   })
 setMethod(f = "sosResult", signature = signature(obj = "OmObservation"),
-          definition = function(obj, coordinates = FALSE) {
+definition = function(obj, coordinates = FALSE) {
             if(coordinates){
               .coords <- sosCoordinates(obj)
               .data <- merge(x = obj@result, y = .coords)
@@ -826,7 +831,7 @@ setMethod(f = "sosResult", signature = signature(obj = "OmObservation"),
             return(obj@result)
           })
 setMethod(f = "sosResult", signature = signature(obj = "OmOM_Observation"),
-          definition = function(obj, coordinates = FALSE) {
+definition = function(obj, coordinates = FALSE) {
             if(coordinates){
               .coords <- sosCoordinates(obj)
               .data <- merge(x = obj@result, y = .coords)
@@ -835,7 +840,7 @@ setMethod(f = "sosResult", signature = signature(obj = "OmOM_Observation"),
             return(obj@result)
           })
 setMethod(f = "sosResult", signature = signature(obj = "OmMeasurement"),
-          definition = function(obj, coordinates = FALSE) {
+definition = function(obj, coordinates = FALSE) {
 
             .obsProp <- sosObservedProperties(obj)
             .value <- obj@result@value
@@ -853,7 +858,7 @@ setMethod(f = "sosResult", signature = signature(obj = "OmMeasurement"),
             else return(.result)
           })
 setMethod(f = "sosResult", signature = signature(obj = "OmObservationProperty"),
-          definition = function(obj, coordinates = FALSE) {
+definition = function(obj, coordinates = FALSE) {
             if(!is.na(obj@href))
               return(c(href = obj@href))
             else if(!is.null(obj@obs))
@@ -861,7 +866,7 @@ setMethod(f = "sosResult", signature = signature(obj = "OmObservationProperty"),
             else return(NA)
           })
 setMethod(f = "sosResult", signature = signature(obj = "OmObservationCollection"),
-          definition = function(obj, coordinates = FALSE, bind = TRUE) {
+definition = function(obj, coordinates = FALSE, bind = TRUE) {
             .l <- lapply(obj@members, sosResult, coordinates = coordinates)
             if(bind)
               .result <- do.call(rbind, .l)
@@ -869,25 +874,25 @@ setMethod(f = "sosResult", signature = signature(obj = "OmObservationCollection"
             return(.result)
           })
 setMethod(f = "sosResult", signature = signature(obj = "list"),
-          def = function(obj, coordinates = FALSE) {
+definition = function(obj, coordinates = FALSE) {
             .l <- lapply(obj, sosResult, coordinates = coordinates)
             .result <- do.call(rbind, .l)
             return(.result)
           })
 setMethod(f = "sosResult", signature = signature(obj = "OwsExceptionReport"),
-          definition = function(obj, coordinates = FALSE) {
+definition = function(obj, coordinates = FALSE) {
             warning("OwsExceptionReport does not have a result set.")
             return(toString(obj))
           })
 setMethod(f = "sosResult", signature = signature(obj = "character"),
-          definition = function(obj, coordinates = FALSE) {
+definition = function(obj, coordinates = FALSE) {
             warning(paste("No processable result given: ", obj))
             return(toString(obj))
           })
 # just returns the data.frame again, allows using the binding facilities of
 # the sosResult(list) function
 setMethod(f = "sosResult", signature = signature(obj = "data.frame"),
-          definition = function(obj, coordinates = FALSE) {
+definition = function(obj, coordinates = FALSE) {
             return(obj)
           })
 
@@ -919,12 +924,12 @@ setMethod(f = "sosCoordinates", signature = signature(obj = "OmObservationProper
             return(c())
           })
 setMethod(f = "sosCoordinates", signature = signature(obj = "OmOM_Observation"),
-          def = function(obj) {
+          definition = function(obj) {
             .coords <- sosCoordinates(obj = obj@featureOfInterest)
             return(.coords)
           })
 setMethod(f = "sosCoordinates", signature = signature(obj = "GmlFeatureCollection"),
-          def = function(obj) {
+          definition = function(obj) {
             .list <- lapply(obj@featureMembers, sosCoordinates)
             .coords <- do.call(rbind, .list)
             return(.coords)
@@ -982,7 +987,7 @@ setMethod(f = "sosCoordinates",
           })
 setMethod(f = "sosCoordinates",
           signature = signature(obj = "list"),
-          definition = function(obj, sos = NULL, verbose = FALSE) {
+definition = function(obj, sos = NULL, verbose = FALSE) {
             if(is.null(sos))
               .list <- lapply(obj, sosCoordinates)
             else .list <- lapply(obj, sosCoordinates, sos = sos, verbose = verbose)
@@ -1048,7 +1053,7 @@ if (!isGeneric("sosName"))
   })
 setMethod(f = "sosName", signature = signature(obj = "list"),
           definition = function(obj) {
-            lapply(obj, sosName)
+            sapply(obj, sosName)
           })
 setMethod(f = "sosName", signature = signature(obj = "SosObservationOffering"),
           definition = function(obj) {
@@ -1146,7 +1151,7 @@ if (!isGeneric("sosEncoders"))
     standardGeneric("sosEncoders")
   })
 setMethod(f = "sosEncoders", signature = signature(sos = "SOS"),
-          def = function(sos) {
+          definition = function(sos) {
             return(sos@encoders)
           })
 if (!isGeneric("sosDataFieldConverters"))
@@ -1166,14 +1171,13 @@ setMethod(f = "sosDataFieldConverters", signature = signature(sos = "SOS_Test"),
 #
 #
 if (!isGeneric("sosUOM"))
-  setGeneric(name = "sosUOM",
-             def = function(obj) {
+  setGeneric(name = "sosUOM", def = function(obj) {
                standardGeneric("sosUOM")
              })
 setMethod(f = "sosUOM",
           signature = c(obj = "list"),
           definition = function(obj) {
-            .crs <- lapply(X = obj, FUN = sosUOM)
+            .crs <- sapply(X = obj, FUN = sosUOM)
             return(.crs)
           }
 )